pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bill-extraction-api"
version = "0.1.0"
description = "FastAPI service for automated bill line-item extraction"
readme = "README.md"
authors = [{ name = "Bajaj Health" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.1",
    "httpx==0.27.0",
    "pydantic==2.8.2",
    "pydantic-settings==2.3.4",
    "python-multipart==0.0.9",
    "pdf2image==1.17.0",
    "pypdf==4.2.0",
    "Pillow==10.4.0",
    "numpy==1.26.4",
    "pandas==2.2.2",
    "rapidocr-onnxruntime>=1.3.19",
    "opencv-python-headless==4.10.0.84",
    "scikit-image==0.24.0",
    "python-dateutil==2.9.0.post0",
    "loguru==0.7.2"
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.2",
    "pytest-asyncio==0.23.8",
    "httpx==0.27.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
